+++ selfdrive/ui/mici/layouts/settings/developer.py
@@ -35,7 +35,7 @@ class DeveloperLayoutMici(NavScroller):
       github_username = ui_state.params.get("GithubUsername") or ""
       dlg = BigInputDialog("enter GitHub username...", github_username, minimum_length=0, confirm_callback=github_username_callback)
       if not system_time_valid():
-        dlg = BigDialog("Please connect to Wi-Fi to fetch your key", "")
+        dlg = BigDialog("", "Please connect to Wi-Fi to fetch your key")
         gui_app.push_widget(dlg)
         return
       gui_app.push_widget(dlg)
